提交客户端代码

+++ Client/Assets/Scripts/Interface/IObj.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public interface IObj
 {
     int        ID             { get; set; }
-    int        GUID           { get; set; }
+    ulong      GUID           { get; set; }
     Vector3    Pos            { get; set; }
     Vector3    Euler          { get; set; }
     Vector3    Scale          { get; set; }
